server/db/database_connection.py
from pymongo import MongoClient
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

# Connect to MongoDB
MONGO_URI = os.getenv("MONGO_URI")
client = MongoClient(MONGO_URI)

# Select database and collection
mydatabase = client["ems"]
mycollection = mydatabase["user"]
