server/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib
bcrypt
pydantic
virtualenv
jwt
email-validator